chore(test): Add tests to explore `failFast` functionality
